requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
transformers==4.44.2
torch>=2.2.0
accelerate>=0.34.2
scikit-learn>=1.3.0
datasets>=2.19.0
evaluate>=0.4.2
pandas>=2.1.0
pyyaml>=6.0.1